Add colored output for packages
Highlight package names in the output by color. If the installed package
presented in the package list for the dependency it will be colored in
green. If the installed package presented in the package list for the
dependency but its version doesn't find in the list it will be colored
in red.